refactor chooseTeam
