+++ lib/chooseTeam.js
@@ -1,8 +1,6 @@
 const rl = require("./rl");
-const logGrid = require("./logGrid");
-const checkWin = require("./checkWin");
 
-const getTeamChoice = () => {
+const promptForTeamChoice = () => {
   return new Promise((resolve) => {
     rl.question("Pick your team. O or X? = ", (choice) => {
       return resolve(choice.toUpperCase());
@@ -10,34 +8,26 @@ const getTeamChoice = () => {
   });
 };
 
-const getValidTeamChoice = (choice) => {
+const isValidTeamChoice = (choice) => {
   if (choice === "O" || choice === "X") {
     return true;
   } else {
+    console.log("Please enter only the letter O or X to choose your team");
     return false;
   }
 };
 
-const setTeam = (store, choice) => {
-  store.userTeam = choice;
-  if (choice === "O") {
-    store.cpuTeam = "X";
-  } else {
-    store.cpuTeam = "O";
-  }
-};
-
-const storeTeamChoice = async (store) => {
-  let choice = await getTeamChoice();
-  if (getValidTeamChoice(choice) === true) {
-    setTeam(store, choice);
-  } else {
-    console.log("Please enter only the letter O or X to choose your team");
-    return storeTeamChoice(store);
+const getTeamChoice = async () => {
+  let x = false;
+  while (x === false) {
+    let team = await promptForTeamChoice();
+    if (isValidTeamChoice(team) === true) {
+      x = true;
+      return team;
+    }
   }
 };
 
 exports.getTeamChoice = getTeamChoice;
-exports.getValidTeamChoice = getValidTeamChoice;
+exports.isValidTeamChoice = isValidTeamChoice;
 exports.setTeam = setTeam;
-exports.storeTeamChoice = storeTeamChoice;
