+++ lib/store.js
@@ -1,0 +1,8 @@
+const setTeam = (store, choice) => {
+  store.userTeam = choice;
+  if (choice === "O") {
+    store.cpuTeam = "X";
+  } else {
+    store.cpuTeam = "O";
+  }
+};
